Helper classes moved to its own package
