+++ src/main/java/hudson/plugins/nested_view/search/ExtendedSearch.java
@@ -1,0 +1,10 @@
+package hudson.plugins.nested_view.search;
+
+import hudson.plugins.nested_view.ProjectWrapper;
+
+// for some reasons, if class implements some interfaces, only methods from interface are available form jelly!
+// thus this artificial interface...
+public interface ExtendedSearch {
+
+    public ProjectWrapper getProject();
+}
