+++ src/main/java/hudson/plugins/nested_view/search/LinkableCandidate.java
@@ -1,4 +1,4 @@
-package hudson.plugins.nested_view;
+package hudson.plugins.nested_view.search;
 
 public class LinkableCandidate {
     private final String plainText;
